Collapse single-re-export wrapper files into their implementations
Delete 7 thin wrapper .ts files that only re-exported from a
-view.tsx counterpart, and rename the -view.tsx files to take
their place directly.
