--- modules/datepicker/datepicker.ts
@@ -1,1 +1,0 @@
-export {DatePicker} from './datepicker-view'
