--- source/views/building-hours/detail/link-table.ts
@@ -1,1 +1,0 @@
-export {LinkTable} from './link-table-view'
