--- source/views/building-hours/detail/schedule-row.ts
@@ -1,1 +1,0 @@
-export {ScheduleRow} from './schedule-row-view'
