+++ source/views/building-hours/detail/schedule-row.tsx
