--- source/views/building-hours/detail/schedule-table.ts
@@ -1,1 +1,0 @@
-export {ScheduleTable} from './schedule-table-view'
