+++ source/views/building-hours/detail/schedule-table.tsx
