--- source/views/directory/detail.ts
@@ -1,1 +1,0 @@
-export {DirectoryDetailView, DetailNavigationOptions} from './detail-view'
