--- source/views/student-orgs/detail.ts
@@ -1,1 +1,0 @@
-export {View, NavigationOptions} from './detail-view'
