+++ source/views/student-orgs/detail.tsx
